+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   This matches the properties returned by the API.
 * Changed type from integer to string for ItemEligibilityResult::itemId and ItemEligibilityResult::transactionId
 * Added missing subdomain property to PostOrder\Error.
+* Added missing discountAmount property to Fulfillment\DeliveryCost.
 
 ## 11.0.0 - 2017-03-20